Move theme hooks to their own file
